Remove helper class V, it is no longer needed

+++ DragSortListView/src/main/java/com/mobeta/android/dslv/DragSortListView.java
@@ -2701,7 +2701,7 @@ public class DragSortListView extends ListView {
 			// dy is change in View position of a list item; i.e. positive dy
 			// means user is scrolling up (list item moves down the screen, remember
 			// y=0 is at top of View).
-			dy = (int) Math.round(mScrollSpeed * dt);
+			dy = Math.round(mScrollSpeed * dt);
 
 			int movePos;
 			if (dy >= 0) {
